* add orgConfig CNS for configuration * remove big file * add amopId * use orgConfig for load contract address

+++ .ci/script/build-ci.sh
@@ -5,7 +5,6 @@ app_xml_config_tpl=${java_source_code_dir}/src/main/resources/fisco.properties.t
 app_xml_config=${java_source_code_dir}/src/main/resources/fisco.properties
 weid_config_tpl=${java_source_code_dir}/src/main/resources/weidentity.properties.tpl
 weid_config=${java_source_code_dir}/src/main/resources/weidentity.properties
-font=${java_source_code_dir}/src/main/resources/NotoSansCJKtc-Regular.ttf
 
 export MYSQL_ADDRESS=${1:-0.0.0.0:3306}
 export MYSQL_DATABASE=${2:-database}
@@ -18,25 +17,23 @@ function modify_config()
 {
     echo "begin to modify sdk config..."
 
-    hash=$(cat hash)
     export FISCO_BCOS_VERSION=${FISCO_BCOS_VERSION}
     export CNS_PROFILE_ACTIVE=${CNS_PROFILE_ACTIVE}
-    export CNS_CONTRACT_FOLLOW=${hash}
-    export CHAIN_ID=${CHAIN_ID}
     
-    MYVARS='${FISCO_BCOS_VERSION}:${CNS_PROFILE_ACTIVE}:${CNS_CONTRACT_FOLLOW}:${CHAIN_ID}'
+    MYVARS='${FISCO_BCOS_VERSION}:${CNS_PROFILE_ACTIVE}'
     envsubst ${MYVARS} < ${app_xml_config_tpl} >${app_xml_config}
     cp ${app_xml_config} ${java_source_code_dir}/src/test/resources/
     # cat $app_xml_config
     
     export ORG_ID=${ORG_ID}
+    export AMOP_ID=${AMOP_ID}
     export MYSQL_ADDRESS=${MYSQL_ADDRESS}
     export MYSQL_DATABASE=${MYSQL_DATABASE}
     export MYSQL_USERNAME=${MYSQL_USERNAME}
     export MYSQL_PASSWORD=${MYSQL_PASSWORD}
     export BLOCKCHIAN_NODE_INFO=${BLOCKCHIAN_NODE_INFO}
     
-    NODEVAR='${ORG_ID}:${MYSQL_ADDRESS}:${MYSQL_DATABASE}:${MYSQL_USERNAME}:${MYSQL_PASSWORD}:${BLOCKCHIAN_NODE_INFO}'
+    NODEVAR='${ORG_ID}:${AMOP_ID}:${MYSQL_ADDRESS}:${MYSQL_DATABASE}:${MYSQL_USERNAME}:${MYSQL_PASSWORD}:${BLOCKCHIAN_NODE_INFO}'
     envsubst ${NODEVAR} < ${weid_config_tpl} >${weid_config}
     cp ${weid_config} ${java_source_code_dir}/src/test/resources/
 
@@ -46,6 +43,8 @@ function modify_config()
     cp ${java_source_code_dir}/.ci/ca.crt ${java_source_code_dir}/src/test/resources
     cp ${java_source_code_dir}/.ci/node.crt ${java_source_code_dir}/src/test/resources
     cp ${java_source_code_dir}/.ci/node.key ${java_source_code_dir}/src/test/resources
+    cp ${java_source_code_dir}/NotoSansCJKtc-Regular.ttf ${java_source_code_dir}/src/main/resources
+    cp ${java_source_code_dir}/NotoSansCJKtc-Regular.ttf ${java_source_code_dir}/src/test/resources
 	cp -r ${java_source_code_dir}/src/main/resources/WeDPR_dynamic_lib ${java_source_code_dir}/src/test/resources
     echo "modify sdk config finished..."
 }
@@ -57,27 +56,24 @@ function gradle_build_sdk()
     cp ${java_source_code_dir}/.ci/ca.crt ${java_source_code_dir}/src/main/resources
     cp ${java_source_code_dir}/.ci/node.crt ${java_source_code_dir}/src/main/resources
     cp ${java_source_code_dir}/.ci/node.key ${java_source_code_dir}/src/main/resources
-    cp ${java_source_code_dir}/dist/lib/NotoSansCJKtc-Regular.ttf ${java_source_code_dir}/src/main/resources
-    cp ${java_source_code_dir}/dist/lib/NotoSansCJKtc-Regular.ttf ${java_source_code_dir}/src/test/resources
-    
+
     export FISCO_BCOS_VERSION="2"
-    export CNS_CONTRACT_FOLLOW=
     export CNS_PROFILE_ACTIVE="ci"
-    export CHAIN_ID=101
     
-    MYVARS='${FISCO_BCOS_VERSION}:${CNS_PROFILE_ACTIVE}:${CNS_CONTRACT_FOLLOW}:${CHAIN_ID}'
+    MYVARS='${FISCO_BCOS_VERSION}:${CNS_PROFILE_ACTIVE}'
     envsubst ${MYVARS} < ${app_xml_config_tpl} >${app_xml_config}
     
     
     content="$NODE_IP"
-    #content="0.0.0.0:8900"
+    # content="0.0.0.0:8902"
     export BLOCKCHIAN_NODE_INFO=${content}
     export ORG_ID="webank_ci"
+    export AMOP_ID="amop_ci"
     echo $MYSQL_ADDRESS
     echo $MYSQL_DATABASE
     echo $MYSQL_USERNAME
     echo $MYSQL_PASSWORD
-    NODEVAR='${ORG_ID}:${MYSQL_ADDRESS}:${MYSQL_DATABASE}:${MYSQL_USERNAME}:${MYSQL_PASSWORD}:${BLOCKCHIAN_NODE_INFO}'
+    NODEVAR='${ORG_ID}:${AMOP_ID}:${MYSQL_ADDRESS}:${MYSQL_DATABASE}:${MYSQL_USERNAME}:${MYSQL_PASSWORD}:${BLOCKCHIAN_NODE_INFO}'
     envsubst ${NODEVAR} < ${weid_config_tpl} >${weid_config}
 
     echo "Begin to compile java code......"
@@ -109,20 +105,32 @@ function deploy_contract()
     do
         CLASSPATH=${CLASSPATH}:${jar_file}
     done
+    
+    chain_id=101
+    privateKey=18602059553666200379844734388296903882431291027699519961839765914892609749994
 
-    java ${JAVA_OPTS} -cp "$CLASSPATH" com.webank.weid.contract.deploy.DeployContract
+    java ${JAVA_OPTS} -cp "$CLASSPATH" com.webank.weid.contract.deploy.DeployContract ${chain_id} ${privateKey}
     echo "contract deployment done."
 }
 
 function  install_font()
 {
-    sudo mkdir -p /usr/share/fonts/custom&&
-    sudo cp ${font} /usr/share/fonts/custom/&&
-    sudo apt install xfonts-utils -y&&
-    sudo mkfontscale&&
-    sudo mkfontdir&&
-    sudo fc-cache -fv
-    echo "font install done."
+    # download NotoSansCJKtc-Regular.ttf
+    if [ ! -f NotoSansCJKtc-Regular.ttf ]; then
+        wget -c https://www.fisco.com.cn/cdn/weevent/weidentity/download/releases/NotoSansCJKtc-Regular.ttf
+    fi
+    
+    if [ -f NotoSansCJKtc-Regular.ttf ]; then
+        sudo mkdir -p /usr/share/fonts/custom&&
+        sudo cp NotoSansCJKtc-Regular.ttf /usr/share/fonts/custom/&&
+        sudo apt install xfonts-utils -y&&
+        sudo mkfontscale&&
+        sudo mkfontdir&&
+        sudo fc-cache -fv
+        echo "font install done."
+    else
+        echo "font install fail."
+    fi
 }
 
 function main()
